refactored how lab.js experiments are structured
