+++ package.json
@@ -281,7 +281,7 @@
     "hazardous": "^0.3.0",
     "history": "^4.7.2",
     "kernelspecs": "^2.0.0",
-    "lab.js": "^20.0.1",
+    "lab.js": "^21.0.0-rc3",
     "lodash": "^4.17.15",
     "lodash.clonedeep": "^4.5.0",
     "mkdirp": "^1.0.4",
